+++ oap-server/server-receiver-plugin/envoy-metrics-receiver-plugin/src/main/java/org/apache/skywalking/oap/server/receiver/envoy/EnvoyMetricReceiverConfig.java
@@ -39,6 +39,8 @@ public class EnvoyMetricReceiverConfig extends ModuleConfig {
     private String alsTCPAnalysis;
     @Getter
     private String k8sServiceNameRule;
+    @Getter
+    private String istioServiceNameRule;
 
     private final ServiceMetaInfoFactory serviceMetaInfoFactory = new ServiceMetaInfoFactoryImpl();
     @Getter